fix: reorganize bus into separate classes - worker and master
