+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "eslint.options" : {
+        "configFile": "C:/ut5/dev/ut-bus/node_modules/ut-tools/eslint/.eslintrc"
+    }
+}
