+++ package.json
@@ -10,7 +10,7 @@
   },
   "devDependencies": {
     "glob": "7.1.2",
-    "ut-tools": "^6.4.6"
+    "ut-tools": "^6.6.0"
   },
   "optionalDependencies": {
     "ut-rpc": "^6.0.0",
